refactor: extract MenuBarIcon geometry and TokenHealthSection session helpers
- Extract star geometry helpers (starPath, brokenStarFragments, drawStroke)
  and NSBezierPath→CGPath extension from MenuBarIcon (502→379 lines) into
  MenuBarIconGeometry.swift (128 lines)
- Extract session detail computation (label parts, tooltip, idle detection,
  time formatting) from TokenHealthSection (413→326 lines) into
  TokenHealthSessionInfo.swift (102 lines) as pure SessionInfoFormatter enum
- All 5 previously-oversized view files now under 400 lines
